HASH flag finished, plus debut norm

+++ ft_itoall.c
@@ -6,7 +6,7 @@
 /*   By: mmanley <mmanley@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/13 15:33:37 by mmanley           #+#    #+#             */
-/*   Updated: 2018/02/23 16:26:44 by mmanley          ###   ########.fr       */
+/*   Updated: 2018/02/23 17:37:07 by mmanley          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ static int		signed_count(long long int nb, int i, int base)
 
 static int		unsigned_signed_count(unsigned long long int nb, int i, int base)
 {
-	if (nb >= base)
+	if (nb >= (unsigned long long int)base)
 		return (rec_nbr_count(nb / base, i + 1, base));
 	return (i);
 }
